+++ frontend/src/components/ArticleSection.tsx
