+++ frontend/src/components/VocabularySection.tsx
